+++ CHANGELOG.md
@@ -12,8 +12,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [2.0.1] - 2021-12-23
 
 ### Fixed
-
 - Fix unpaid events with no attendees being categorized as paid
+- Fix invalid event progress warning indicator with no paid attendees
 
 ## [2.0.0] - 2021-12-22